BUGFIX: mmap(2) requires the 'offset' argument to be the index         of the 'nth' *mapping* times the page-size.
However, IOPORT regions are not counted as 'mappings' since
it is not possible to map such regions.

Therefore, the index of a mapping is not necessarily
equal to the BAR index.

The committed fix subtracts the number of IOPORT regions
below a given BAR from the BAR index yielding the mapping
index.

The fix also rejects an attempt to map a IO port region.